Find the Sum Of the Element From Index L to R  Before Update & After Update For Given Array Elements.
Given an array arr[] of n integers. The task is to do the following operations:  

Add a value X to all the element  from index A to B where 0 ≤ A ≤ B ≤ N-1.
Find the sum of the element from index L to R  before update and after update for given array elements.

Example: 
Input: n=6,arr[] = {1, 3, 5, 7, 9, 11}, L = 1, R = 3, A = 1, B = 5, X = 10 
Output:             
Sum of values in given range before  = 15 
Updated sum of values in given range = 45 

Explanation: 
Sum of values in the range 1 to 3 before update is 3 + 5 + 7 = 15. 
arr[] after adding 10 from index 1 to 5 is arr[] = {1, 13, 15, 17, 19, 21} 
Sum of values in the range 1 to 3 after update is 13 + 15 + 17 = 45.